Minor clean
cache Generation value, store arceus stuff as ushort

+++ PKHeX.Core/Legality/Tables/Tables.cs
@@ -308,8 +308,8 @@ namespace PKHeX.Core
             797, 798, 799, 803, 804, 805, 806,
         };
 
-        public static readonly int[] Arceus_Plate = {303, 306, 304, 305, 309, 308, 310, 313, 298, 299, 301, 300, 307, 302, 311, 312, 644};
-        public static readonly int[] Arceus_ZCrystal = {782, 785, 783, 784, 788, 787, 789, 792, 777, 778, 780, 779, 786, 781, 790, 791, 793};
+        public static readonly ushort[] Arceus_Plate = {303, 306, 304, 305, 309, 308, 310, 313, 298, 299, 301, 300, 307, 302, 311, 312, 644};
+        public static readonly ushort[] Arceus_ZCrystal = {782, 785, 783, 784, 788, 787, 789, 792, 777, 778, 780, 779, 786, 781, 790, 791, 793};
 
         internal static readonly int[] BabyEvolutionWithMove =
         {
